Cannot get my own forked-repo GitHub Issues in VS Code Fixes #3573
