+++ src/issues/currentIssue.ts
@@ -8,13 +8,13 @@ import { Repository } from '../api/api';
 import { Remote } from '../common/remote';
 import { FolderRepositoryManager, PullRequestDefaults } from '../github/folderRepositoryManager';
 import { IssueModel } from '../github/issueModel';
+import { variableSubstitution } from '../github/utils';
 import { IssueState, StateManager } from './stateManager';
 import {
 	BRANCH_CONFIGURATION,
 	BRANCH_NAME_CONFIGURATION,
 	ISSUES_CONFIGURATION,
-	SCM_MESSAGE_CONFIGURATION,
-	variableSubstitution,
+	SCM_MESSAGE_CONFIGURATION
 } from './util';
 
 export class CurrentIssue {
