+++ src/issues/issueHoverProvider.ts
@@ -7,12 +7,12 @@ import * as vscode from 'vscode';
 import { ITelemetry } from '../common/telemetry';
 import { FolderRepositoryManager } from '../github/folderRepositoryManager';
 import { RepositoriesManager } from '../github/repositoriesManager';
+import { ParsedIssue } from '../github/utils';
 import { StateManager } from './stateManager';
 import {
 	getIssue,
 	ISSUE_OR_URL_EXPRESSION,
 	issueMarkdown,
-	ParsedIssue,
 	parseIssueExpressionOutput,
 	shouldShowHover,
 } from './util';
