+++ src/issues/stateManager.ts
@@ -19,14 +19,13 @@ import { IAccount } from '../github/interface';
 import { IssueModel } from '../github/issueModel';
 import { MilestoneModel } from '../github/milestoneModel';
 import { RepositoriesManager } from '../github/repositoriesManager';
+import { getIssueNumberLabel, variableSubstitution } from '../github/utils';
 import { CurrentIssue } from './currentIssue';
 import {
 	BRANCH_CONFIGURATION,
 	DEFAULT_QUERY_CONFIGURATION,
-	getIssueNumberLabel,
 	ISSUES_CONFIGURATION,
 	QUERIES_CONFIGURATION,
-	variableSubstitution,
 } from './util';
 
 // TODO: make exclude from date words configurable
@@ -316,7 +315,8 @@ export class StateManager {
 				}
 				items = this.setIssues(
 					folderManager,
-					await variableSubstitution(query.query, undefined, defaults, user),
+					// Do not resolve pull request defaults as they will get resolved in the query later per repository
+					await variableSubstitution(query.query, undefined, undefined, user),
 				);
 			}
 			singleRepoState.issueCollection.set(query.label, items);
@@ -328,7 +328,7 @@ export class StateManager {
 
 	private setIssues(folderManager: FolderRepositoryManager, query: string): Promise<IssueItem[]> {
 		return new Promise(async resolve => {
-			const issues = await folderManager.getIssues({ fetchNextPage: false }, query);
+			const issues = await folderManager.getIssues({ fetchNextPage: false, fetchOnePagePerRepo: true }, query);
 			this._onDidChangeIssueData.fire();
 			resolve(
 				issues.items.map(item => {
